Add the incomplete splitscreen functionality that I developed
This works by alternating every frame between rendering to
the left half of the screen and the right half of the screen.

Currently, the two screens are (almost) identical; for the
two screens to each track a different player, the camera
likewise needs to alternate every frame between its
position tracking player one and its position tracking
player two.